feat: moved metrics to private endpoint
