+++ terraform/ecs/variables.tf
@@ -18,6 +18,10 @@ variable "port" {
   type = number
 }
 
+variable "private_port" {
+  type = number
+}
+
 variable "acm_certificate_arn" {
   type = string
 }
