+++ tests/functional/connection/infura.rs
@@ -27,24 +27,6 @@ async fn health_check(ctx: &mut ServerContext) {
 
 #[test_context(ServerContext)]
 #[tokio::test]
-async fn metrics_check(ctx: &mut ServerContext) {
-    let addr = format!("htpps://{}/metrics", ctx.server.public_addr);
-
-    let client = Client::builder().build::<_, hyper::Body>(HttpsConnector::new());
-
-    let request = Request::builder()
-        .method(Method::GET)
-        .uri(addr)
-        .body(Body::default())
-        .unwrap();
-
-    let response = client.request(request).await.unwrap();
-
-    assert_eq!(response.status(), http::StatusCode::OK)
-}
-
-#[test_context(ServerContext)]
-#[tokio::test]
 async fn eip155_1_mainnet_infura(ctx: &mut ServerContext) {
     let addr = format!(
         "http://{}/v1?projectId={}&chainId=",
